* set default LED brightness to 100

* testing LEDs


* add one-click demos; restructure menu

* restructure menus, fid desktop icons

* remove kivy icon from default desktop layout

* change Qiskit default to 0.39; update Qoffee icon

* adding Qiskit 0.39

* remove qiskit version from requirements.txt

* add Qiskit 0.41

* remove qiskit-widgets from requirements.txt

* add green button

* add luma setup and configure green button

* workaround to force conan 1.x for Qiskit 0.41

* force conan 1.x for qiskit 0.39

* force conan 1.x for qiskit 039, 039

* force conan 1.x

* fix qiskit 0.41 install

* add Qiskit 0.42

* add Qiskit 0.42

* merge master to dev 10 (fractals, qiskit 0.43, green button, various fixes) (#146)

* merge dev10 to master: remove kivy icon from default desktop layout  (#140)

* remove kivy icon from default desktop layout

* Quantum Fractal Demo (#141)

* add one-click demos; restructure menu

* restructure menus, fid desktop icons

* merge dev10 to master: remove kivy icon from default desktop layout  (#140)

* remove kivy icon from default desktop layout

* Create main.yml

* Update main.yml

* Delete .github/workflows directory

* Added packages used for Fractal Demo

* Included PyCharm project settings to .gitignore


* Delete .idea directory

* Delete test.txt

* Add new Qiskit versions, add luma for LED matrix display, add green button (#143)

* set default LED brightness to 100

* testing LEDs

* fix LED invert

* fix LED invert

* fix LED invert

* fix LED invert

* fix LED invert

* fix LED invert

* fix LED invert

* add one-click demos; restructure menu

* restructure menus, fid desktop icons

* remove kivy icon from default desktop layout

* change Qiskit default to 0.39; update Qoffee icon

* adding Qiskit 0.39

* remove qiskit version from requirements.txt

* add Qiskit 0.41

* remove qiskit-widgets from requirements.txt

* add green button

* add luma setup and configure green button

* workaround to force conan 1.x for Qiskit 0.41

* force conan 1.x for qiskit 0.39

* force conan 1.x for qiskit 039, 039

* force conan 1.x

* fix qiskit 0.41 install

* add Qiskit 0.42

* add Qiskit 0.42

* add Qiskit 0.43 and whl file for quantum-widgets

* merge dev 10 (Qiskit 0.42, green button, various fixes) (#145)

* set default LED brightness to 100

* testing LEDs

* add one-click demos; restructure menu

* restructure menus, fix desktop icons

* remove kivy icon from default desktop layout

* change Qiskit default to 0.39; update Qoffee icon

* adding Qiskit 0.39

* remove qiskit version from requirements.txt

* add Qiskit 0.41

* remove qiskit-widgets from requirements.txt

* add green button

* add luma setup and configure green button

* workaround to force conan 1.x for Qiskit 0.41

* force conan 1.x for qiskit 0.39

* force conan 1.x for qiskit 039, 039

* force conan 1.x

* fix qiskit 0.41 install

* add Qiskit 0.42

* add Qiskit 0.42

---------


* reorder auto install procedure; use local quantum-widgets whl

* moved RasQ-init.sh from gist to GH repo

* update RasQ-init.sh for Bookworm 64-bit mode

* initial support for Bookworm 64-bit OS and Qiskit 1.0

* add OS_VERSION to environment

* typo

* typo

* typo

* typo

* load env

* install Qiskit 1.0.x on bookworm 64-bit

* update on-click install for Qiskit 1.0 and bookworm

* update

* add Qiskit 0.44 and make default

* use local quantum-widgets whl

---------

+++ RasQ-init.sh
@@ -1,8 +1,7 @@
 #!/bin/bash
 #
 # usage (in RPi terminal):
-# pip3 install getgist
-# .local/bin/getgist -y JanLahmann RasQ-init.sh
+# wget https://github.com/JanLahmann/RasQberry/raw/master/RasQ-init.sh
 # . ./RasQ-init.sh
 # also use the previous command to start the rasqberry-config tool again,
 # or   sudo rasqberry-config
@@ -61,18 +60,21 @@ fi
 #[ ! -d ~/Desktop ] && mkdir ~/Desktop
 #[ -d RasQberry/desktop-icons ] && cp RasQberry/desktop-icons/*.desktop ~/Desktop
 
-# force 32 bit kernel  being used
+# force 32 bit kernel  being used on Bullseye OS. Stay with 64-bit on Bookworm and later OS.
 
-if ! grep -q -E "^arm_64bit=0" /boot/config.txt; then
-  echo; echo;
-  echo "**********************************************************************"
-  echo "/boot/config.txt will be modified to ensure 32 bit kernel being used"
-  sudo -H -- sh -c 'echo "\n# ensure 32 bit kernel being used\narm_64bit=0\n" >> /boot/config.txt'
-  echo "rebooting after modification of /boot/config.txt"
-  echo "After reboot, please login and restart the script RasQ-init.sh"
-  echo "**********************************************************************"
-  echo;
-  sudo reboot
+if grep -q -E "bullseye" /etc/os-release ; then  
+  if ! grep -q -E "^arm_64bit=0" /boot/config.txt; then
+    echo; echo;
+    echo "**********************************************************************"
+    echo "bookworm OS detected. Forcing 32-bit mode."
+    echo "/boot/config.txt will be modified to ensure 32 bit kernel being used"
+    sudo -H -- sh -c 'echo "\n# ensure 32 bit kernel being used\narm_64bit=0\n" >> /boot/config.txt'
+    echo "rebooting after modification of /boot/config.txt"
+    echo "After reboot, please login and restart the script RasQ-init.sh"
+    echo "**********************************************************************"
+    echo;
+    sudo reboot
+  fi
 fi
 
 sudo rasqberry-config
